% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsR.R
\name{browser}
\alias{browser}
\title{browser}
\description{
browser

browser
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{browser$new()}}
\item \href{#method-list_tabs}{\code{browser$list_tabs()}}
\item \href{#method-execute}{\code{browser$execute()}}
\item \href{#method-clone}{\code{browser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{browser$new(debug_port = 9222)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_tabs"></a>}}
\subsection{Method \code{list_tabs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{browser$list_tabs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\subsection{Method \code{execute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{browser$execute(title = "", script = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{browser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
