% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{browser_launch}
\alias{browser_launch}
\title{Launches a Playwright browser instance}
\usage{
browser_launch(
  browser = c("chromium", "firefox", "webkit"),
  headless = TRUE,
  timeout = 30000,
  user_data_dir = NULL,
  channel = NULL,
  user_agent = NULL
)
}
\arguments{
\item{headless}{Whether to run the browser in headless mode. Default is \code{TRUE}.}

\item{timeout}{Maximum time in milliseconds to wait for the browser instance to start. Default is 30000 (30 seconds).
Pass 0 to disable timeout.}

\item{executable_path}{Path to a browser executable to run instead of the bundled one. If \code{NULL},
the bundled browser will be used. Default is \code{NULL}.}

\item{args}{Additional arguments to pass to the browser instance. Default is \code{NULL}.}

\item{proxy}{Proxy settings for all requests. Default is \code{NULL}.}

\item{slow_mo}{Slows down Playwright operations by the specified amount of milliseconds. Useful for debugging.
Default is \code{NULL}.}
}
\value{
A tibble containing the ID of the launched browser.
}
\description{
This function launches a Playwright browser instance and returns a tibble
containing the ID of the browser. The browser instance can be used to create
new browser contexts and perform various actions.
}
\examples{
browser <- browser_launch()

}
