% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{check}
\alias{check}
\title{Page class: Check a checkbox}
\usage{
check(page_df, selector = NULL)
}
\arguments{
\item{page_df}{A data frame containing page information, including the \code{page_id} column.
This data frame is used to identify the page object where the checkbox should be checked.}

\item{selector}{(Optional) A character string representing the selector of the checkbox.
If provided, the checkbox will be located using the selector. If not provided, the \code{elem_id} column from \code{page_df} will be used to reference a previously located element.}
}
\value{
The updated \code{page_df} data frame.
}
\description{
The \code{check} function checks a checkbox element. It can either locate the checkbox using a provided selector or by referencing an element previously located.
}
\examples{
# Create a page_df data frame
page_df <- data.frame(page_id = "my_page")

# Check a checkbox using a selector
check(page_df, selector = "#myCheckbox")

# Check a checkbox using a previously located element
page_df <- get_by_role(page_df, "checkbox")
check(page_df)

}
