% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{close_page}
\alias{close_page}
\title{Page class: Close a page object}
\usage{
close_page(page_df)
}
\arguments{
\item{page_df}{A data frame containing page information, including the \code{page_id} column.
This data frame is used to identify the page object that needs to be closed.}
}
\description{
The \code{close_page} function closes a page object within a web browser session using the provided \code{page_df} data frame.
}
\examples{
# Create a page_df data frame
page_df <- data.frame(page_id = "my_page")

# Close the page using the page_df
close_page(page_df)


}
