% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{dblclick}
\alias{dblclick}
\title{Page class: Double-click an element}
\usage{
dblclick(page_df, selector = NULL)
}
\arguments{
\item{page_df}{A data frame containing page information, including the \code{page_id} column.
This data frame is used to identify the page object where the element should be double-clicked.}

\item{selector}{(Optional) A character string representing the selector of the element.
If provided, the element will be located using the selector. If not provided, the \code{elem_id} column from \code{page_df} will be used to reference a previously located element.}
}
\value{
The updated \code{page_df} data frame.
}
\description{
The \code{dblclick} function performs a double click on an element. It can either locate the element using a provided selector or by referencing an element previously located.
}
\examples{
# Create a page_df data frame
page_df <- data.frame(page_id = "my_page")

# Double-click an element using a selector
dblclick(page_df, selector = "#myElement")

# Double-click an element using a previously located element
page_df <- get_by_text(page_df, "Click Me")
dblclick(page_df)

}
