% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{drag_to}
\alias{drag_to}
\title{Page class: Drag and drop an element}
\usage{
drag_to(page_df, to_elem_id, selector = NULL)
}
\arguments{
\item{page_df}{A data frame containing page information, including the \code{page_id} column.
This data frame is used to identify the page object where the drag and drop operation should be performed.}

\item{to_elem_id}{A character string representing the \code{elem_id} of the target element.}

\item{selector}{(Optional) A character string representing the selector of the source element.
If provided, the source element will be located using the selector. If not provided, the \code{elem_id} column from \code{page_df} will be used to reference a previously located element.}
}
\value{
The updated \code{page_df} data frame.
}
\description{
The \code{drag_to} function performs a drag and drop operation from the source element to the target element. It can either locate the elements using provided selectors or by referencing elements previously located.
}
\examples{
# Create a page_df data frame
page_df <- data.frame(page_id = "my_page")

# Drag and drop an element using selectors
drag_to(page_df, to_elem_id = "myTargetElement", selector = "#mySourceElement")

# Drag and drop an element using previously located elements
page_df <- get_by_text(page_df, "Drag Me")
page_df <- get_by_text(page_df, "Drop Here")
drag_to(page_df, to_elem_id = page_df$elem_id[2])

}
