% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_utils.R
\name{execute_script}
\alias{execute_script}
\title{Page class: Execute JavaScript code on the page}
\usage{
execute_script(page_df, script)
}
\arguments{
\item{page_df}{A data frame containing page information, including the \code{page_id} column.
This data frame is used to identify the page object on which the JavaScript code should be executed.}

\item{script}{A character string representing the JavaScript code to execute on the page.}
}
\value{
The result of executing the JavaScript code on the page.
}
\description{
The \code{execute_script} function executes the specified JavaScript code on the page object within a web browser session.
}
\examples{
# Create a page_df data frame
page_df <- data.frame(page_id = "my_page")

# Execute JavaScript code on the page using the page_df
execute_script(page_df, "alert('Hello, World!');")


}
