% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{fill}
\alias{fill}
\title{Page class: Fill input field}
\usage{
fill(page_df, text, selector = NULL)
}
\arguments{
\item{page_df}{A data frame containing page information, including the \code{page_id} column.
This data frame is used to identify the page object where the input field should be filled.}

\item{text}{A character string representing the text to fill in the input field.}

\item{selector}{(Optional) A character string representing the selector of the input field.
If provided, the input field will be located using the selector. If not provided, the \code{elem_id} column from \code{page_df} will be used to reference a previously located element.}
}
\value{
The updated \code{page_df} data frame.
}
\description{
The \code{fill} function fills an input field with the specified text. It can either locate the input field using a provided selector or by referencing an element previously located.
}
\examples{
# Create a page_df data frame
page_df <- data.frame(page_id = "my_page")

# Fill an input field using a selector
fill(page_df, "Hello, World!", selector = "#myInput")

# Fill an input field using a previously located element
page_df <- get_by_role(page_df, "textbox")
fill(page_df, "Hello, World!")
\%>\%
}
