% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{focus}
\alias{focus}
\title{Page class: Set focus on an element}
\usage{
focus(page_df, selector = NULL)
}
\arguments{
\item{page_df}{A data frame containing page information, including the \code{page_id} column.
This data frame is used to identify the page object where the focus should be set.}

\item{selector}{(Optional) A character string representing the selector of the element.
If provided, the element will be located using the selector. If not provided, the \code{elem_id} column from \code{page_df} will be used to reference a previously located element.}
}
\value{
The updated \code{page_df} data frame.
}
\description{
The \code{focus} function sets focus on an element. It can either locate the element using a provided selector or by referencing an element previously located.
}
\examples{
# Create a page_df data frame
page_df <- data.frame(page_id = "my_page")

# Set focus on an element using a selector
focus(page_df, selector = "#myElement")

# Set focus on an element using a previously located element
page_df <- get_by_role(page_df, "textbox")
focus(page_df)

}
