% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_utils.R
\name{get_content}
\alias{get_content}
\title{Page class: Get the page content}
\usage{
get_content(page_df)
}
\arguments{
\item{page_df}{A data frame containing page information, including the \code{page_id} column.
This data frame is used to identify the page object from which the content should be retrieved.}
}
\value{
An XML document representing the content of the page.
}
\description{
The \code{get_content} function retrieves the content of the page object within a web browser session.
}
\examples{
# Create a page_df data frame
page_df <- data.frame(page_id = "my_page")

# Get the page content using the page_df
get_content(page_df)

}
