% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_utils.R
\name{get_title}
\alias{get_title}
\title{Page class: Get the page title}
\usage{
get_title(page_df)
}
\arguments{
\item{page_df}{A data frame containing page information, including the \code{page_id} column.
This data frame is used to identify the page object from which the title should be retrieved.}
}
\value{
A character string representing the title of the page.
}
\description{
The \code{get_title} function retrieves the title of the page object within a web browser session.
}
\examples{
# Create a page_df data frame
page_df <- data.frame(page_id = "my_page")

# Get the page title using the page_df
get_title(page_df)


}
