% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_utils.R
\name{get_url}
\alias{get_url}
\title{Page class: Get the current page URL}
\usage{
get_url(page_df)
}
\arguments{
\item{page_df}{A data frame containing page information, including the \code{page_id} column.
This data frame is used to identify the page object from which the URL should be retrieved.}
}
\value{
A character string representing the current URL of the page.
}
\description{
The \code{get_url} function retrieves the current URL of the page object within a web browser session.
}
\examples{
# Create a page_df data frame
page_df <- data.frame(page_id = "my_page")

# Get the current page URL using the page_df
get_url(page_df)


}
