% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav_utils.R
\name{go_back}
\alias{go_back}
\title{Page class: Go back in browser history}
\usage{
go_back(page_df)
}
\arguments{
\item{page_df}{A data frame containing page information, including the \code{page_id} column.
This data frame is used to identify the page object for which the navigation should occur.}
}
\value{
The updated \code{page_df} data frame.
}
\description{
The \code{go_back} function navigates the page object within a web browser session to the previous page in the browser history.
}
\examples{
# Create a page_df data frame
page_df <- data.frame(page_id = "my_page")

# Go back in browser history using the page_df
go_back(page_df)


}
