% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav_utils.R
\name{go_forward}
\alias{go_forward}
\title{Page class: Go forward in browser history}
\usage{
go_forward(page_df)
}
\arguments{
\item{page_df}{A data frame containing page information, including the \code{page_id} column.
This data frame is used to identify the page object for which the navigation should occur.}
}
\value{
The updated \code{page_df} data frame.
}
\description{
The \code{go_forward} function navigates the page object within a web browser session to the next page in the browser history.
}
\examples{
# Create a page_df data frame
page_df <- data.frame(page_id = "my_page")

# Go forward in browser history using the page_df
go_forward(page_df)


}
