% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav_utils.R
\name{goto}
\alias{goto}
\title{Page class: Go to a URL}
\usage{
goto(page_df, url)
}
\arguments{
\item{page_df}{A data frame containing page information, including the \code{page_id} column.
This data frame is used to identify the page object where the navigation will occur.}

\item{url}{A character string representing the URL to navigate to.}
}
\value{
The updated \code{page_df} data frame.
}
\description{
The \code{goto} function navigates the page object within a web browser session to the specified URL.
}
\examples{
# Create a page_df data frame
page_df <- data.frame(page_id = "my_page")

# Navigate to a URL using the page_df
goto(page_df, "https://www.example.com")


}
