% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{press}
\alias{press}
\title{Page class: Press a key}
\usage{
press(page_df, key, selector = NULL)
}
\arguments{
\item{page_df}{A data frame containing page information, including the \code{page_id} column.
This data frame is used to identify the page object where the key press should be simulated.}

\item{key}{A character string representing the key to press.}

\item{selector}{(Optional) A character string representing the selector of the element.
If provided, the element will be located using the selector. If not provided, the \code{elem_id} column from \code{page_df} will be used to reference a previously located element.}
}
\value{
The updated \code{page_df} data frame.
}
\description{
The \code{press} function simulates pressing a key on the keyboard. It can either locate the element using a provided selector or by referencing an element previously located.
}
\examples{
# Create a page_df data frame
page_df <- data.frame(page_id = "my_page")

# Press a key using a selector
press(page_df, "Enter", selector = "#myElement")

# Press a key using a previously located element
page_df <- get_by_role(page_df, "textbox")
press(page_df, "Enter")

}
