% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_init.R
\name{pw_init}
\alias{pw_init}
\title{Initialize Playwright with optional XVFB}
\usage{
pw_init(use_xvfb = TRUE)
}
\arguments{
\item{use_xvfb}{A boolean value indicating whether to start XVFB using XVFBWrapper.
If set to TRUE and XVFB has not been started already, XVFB will be started before initializing Playwright.
If set to FALSE or if XVFB has already been started, XVFB will not be started.}
}
\description{
The \code{pw_init} function initializes Playwright and optionally starts XVFB if it has not been started already.
}
\examples{
# Initialize Playwright without XVFB
pw_init(use_xvfb = FALSE)

# Initialize Playwright with XVFB
pw_init(use_xvfb = TRUE)

}
