% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav_utils.R
\name{reload}
\alias{reload}
\title{Page class: Reload the current page}
\usage{
reload(page_df)
}
\arguments{
\item{page_df}{A data frame containing page information, including the \code{page_id} column.
This data frame is used to identify the page object that needs to be reloaded.}
}
\value{
The updated \code{page_df} data frame.
}
\description{
The \code{reload} function reloads the current page within a web browser session.
}
\examples{
# Create a page_df data frame
page_df <- data.frame(page_id = "my_page")

# Reload the page using the page_df
reload(page_df)


}
