% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_utils.R
\name{screenshot}
\alias{screenshot}
\title{Page class: Capture a screenshot}
\usage{
screenshot(page_df, path)
}
\arguments{
\item{page_df}{A data frame containing page information, including the \code{page_id} column.
This data frame is used to identify the page object for which the screenshot should be captured.}

\item{path}{A character string representing the path where the screenshot should be saved.}
}
\value{
The updated \code{page_df} data frame.
}
\description{
The \code{screenshot} function captures a screenshot of the current page within a web browser session and saves it to the specified \code{path}.
}
\examples{
# Create a page_df data frame
page_df <- data.frame(page_id = "my_page")

# Capture a screenshot using the page_df and save it to the specified path
screenshot(page_df, "path/to/save/screenshot.png")


}
