% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{scroll}
\alias{scroll}
\title{Scroll on a page}
\usage{
scroll(page_df, location = "scrollHeight")
}
\arguments{
\item{page_df}{A data frame containing page information, including the \code{page_id} column.
This data frame is used to identify the page object where the drag and drop operation should be performed.}

\item{location}{A location string (see examples)}
}
\description{
Scroll on a page
}
\examples{
\dontrun{
# scroll to bottom of the page
scroll(page, location = "scrollHeight")

# scroll to top of the page
scroll(page, location = "scrollTop")

# scroll 100px down from the top of the page
scroll(page, location = "scrollTop - 100")
}
}
