% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{select_option}
\alias{select_option}
\title{Page class: Select an option from a dropdown}
\usage{
select_option(page_df, option, selector = NULL)
}
\arguments{
\item{page_df}{A data frame containing page information, including the \code{page_id} column.
This data frame is used to identify the page object where the dropdown should be selected.}

\item{option}{A character string representing the option to select from the dropdown.}

\item{selector}{(Optional) A character string representing the selector of the dropdown.
If provided, the dropdown will be located using the selector. If not provided, the \code{elem_id} column from \code{page_df} will be used to reference a previously located element.}
}
\value{
The updated \code{page_df} data frame.
}
\description{
The \code{select_option} function selects an option from a dropdown element. It can either locate the dropdown using a provided selector or by referencing an element previously located.
}
\examples{
# Create a page_df data frame
page_df <- data.frame(page_id = "my_page")

# Select an option from a dropdown using a selector
select_option(page_df, "Option 1", selector = "#myDropdown")

# Select an option from a dropdown using a previously located element
page_df <- get_by_role(page_df, "combobox")
select_option(page_df, "Option 2")

}
