% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{type}
\alias{type}
\title{Page class: Type text into an element}
\usage{
type(page_df, text, selector = NULL)
}
\arguments{
\item{page_df}{A data frame containing page information, including the \code{page_id} column.
This data frame is used to identify the page object where the text should be typed.}

\item{text}{A character string representing the text to type.}

\item{selector}{(Optional) A character string representing the selector of the element.
If provided, the element will be located using the selector. If not provided, the \code{elem_id} column from \code{page_df} will be used to reference a previously located element.}
}
\value{
The updated \code{page_df} data frame.
}
\description{
The \code{type} function types text into an element. It can either locate the element using a provided selector or by referencing an element previously located.
}
\examples{
# Create a page_df data frame
page_df <- data.frame(page_id = "my_page")

# Type text into an element using a selector
type(page_df, "Hello, World!", selector = "#myElement")

# Type text into an element using a previously located element
page_df <- get_by_role(page_df, "textbox")
type(page_df, "Hello, World!")

}
